Initial Code Commit
This is the code for the Enterprise. It controls both the Linear
Actuator and Motor with an arduino, 4-channel relay, battery, and a few
buttons. It is based on the sample Arduino code Debounce. It works
pretty well but the Linear Actuator code will run without the button
being pressed sometimes. This might be because the HIGH of the Motor
pin might be bleeding into the Linear Actuator pin. The cycle for the
Linear Actuator will have to be updated and maybe edited so it doesn’t
freeze the code. This can be done by using time lapsed and having that
updated. I might have to look into it more.