Debugging and tweaking
Changed a few pins. Increase Button Debounce Delay. First iteration of
printed board the buttons don’t seem to work equally well. I found out
the best results are with the Linear Actuator on the first button and
Motor Button on the second one. Also the wires are touchy on the 2nd
one. I’m consider resoldering the headers and reprinting so VCC and
ground doesn’t have to go through the button.