Add Button Debounce to Motor Method
Added a button debounce to motor Method. Changed the pin 13 to 12 so
that during the code upload. The pin doesn’t pulse. Added volatile to a
few variables that interact with the interrupt. That lets the compiler
know that the variable can change anytime without any calls around the
code.