PWM in Motor
Added a new sketch where in the Motor function it uses a PWM signal to
control the relay. In Beta Mode. Switched pins around in other sketch.
Tried to change the PWM frequency. It kinda worked. The relay would
shut off with the PWM signal. Thinking about that now that might be a
safety feature of some sort. The Relay is Active LOW.