Debugging Interrupt Code with Tai
Heavily edited Interrupt code to try and debug errors. A lot of the
debounce procedure what edited and simplified. Serial printing was
turned off to no avail. I’m going to try to take off the power on the
empty relay. Also the interrupts worked fine without a motor and linear
actuator on the relay so looking into that will prob be the solution.